+++ src/routers/auth.router.js
@@ -11,44 +11,46 @@ import {
   HASH_SALT_ROUNDS,
 } from '../constants/auth.constant.js';
 import { ACCESS_TOKEN_SECRET } from '../constants/env.constant.js';
+import { signUpController } from '../controllers/auth.controller.js';
 
 const authRouter = express.Router();
-// 회원가입 API
-authRouter.post('/sign-up', signUpValidator, async (req, res, next) => {
-  try {
-    const { email, password, name } = req.body;
+// // 회원가입 API
+authRouter.post('/sign-up', signUpValidator, signUpController);
+// authRouter.post('/sign-up', signUpValidator, async (req, res, next) => {
+//   try {
+//     const { email, password, name } = req.body;
 
-    const existedUser = await prisma.user.findUnique({ where: { email } });
+//     const existedUser = await prisma.user.findUnique({ where: { email } });
 
-    // 이메일이 중복된 경우
-    if (existedUser) {
-      return res.status(HTTP_STATUS.CONFLICT).json({
-        status: HTTP_STATUS.CONFLICT,
-        message: MESSAGES.AUTH.COMMON.EMAIL.DUPLICATED,
-      });
-    }
+//     // 이메일이 중복된 경우
+//     if (existedUser) {
+//       return res.status(HTTP_STATUS.CONFLICT).json({
+//         status: HTTP_STATUS.CONFLICT,
+//         message: MESSAGES.AUTH.COMMON.EMAIL.DUPLICATED,
+//       });
+//     }
 
-    const hashedPassword = bcrypt.hashSync(password, HASH_SALT_ROUNDS);
+//     const hashedPassword = bcrypt.hashSync(password, HASH_SALT_ROUNDS);
 
-    const data = await prisma.user.create({
-      data: {
-        email,
-        password: hashedPassword,
-        name,
-      },
-    });
+//     const data = await prisma.user.create({
+//       data: {
+//         email,
+//         password: hashedPassword,
+//         name,
+//       },
+//     });
 
-    data.password = undefined;
+//     data.password = undefined;
 
-    return res.status(HTTP_STATUS.CREATED).json({
-      status: HTTP_STATUS.CREATED,
-      message: MESSAGES.AUTH.SIGN_UP.SUCCEED,
-      data,
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+//     return res.status(HTTP_STATUS.CREATED).json({
+//       status: HTTP_STATUS.CREATED,
+//       message: MESSAGES.AUTH.SIGN_UP.SUCCEED,
+//       data,
+//     });
+//   } catch (error) {
+//     next(error);
+//   }
+// });
 // 로그인 API
 authRouter.post('/sign-in', signInValidator, async (req, res, next) => {
   try {
